+++ CHANGELOG.md
@@ -21,6 +21,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <hr />
 
+## `0.3.0` (2023-Dec-20)
+
+### Added
+- Optional database persistence (via `--persist <path>` flag)
+
+### Changed
+- Improve entity stubbing to initialize nullable fields
+
+### Fixed
+- API error responses were not fully defined
+- Set `NODE_ENV` to `production` when bundling for NPM
+- Database reset was referencing mutated values (not original)
+
+<hr />
+
 ## `0.2.0` (2023-Dec-19)
 
 ### Added